+++ src/index.ts
@@ -337,6 +337,8 @@ interface AuthLobbyClassifiedInfo {
 async function importClassifiedInfo(
   classifiedInfo: AuthLobbyClassifiedInfo
 ): Promise<void> {
+
+  console.log(classifiedInfo)
   // Extract session key and its iv
   const rawSessionKey = await crypto.keystore.decrypt(classifiedInfo.sessionKey)